#2135: moved components into files and cleaned up code
