+++ src/frontend/src/pages/DesignReviewSummaryModal.tsx
@@ -1,6 +1,6 @@
-import { ChangeRequestType, DesignReview, User, WbsNumber, wbsPipe } from 'shared';
+import { DesignReview, DesignReviewStatus } from 'shared';
 import NERModal from '../components/NERModal';
-import { Box, Button, FormControlLabel, Grid, IconButton, Link, TextField, Typography } from '@mui/material';
+import { Box, Button, FormControlLabel, Grid, IconButton, Typography } from '@mui/material';
 import AccessTimeIcon from '@mui/icons-material/AccessTime';
 import LocationOnIcon from '@mui/icons-material/LocationOn';
 import DescriptionIcon from '@mui/icons-material/Description';
@@ -8,159 +8,11 @@ import VideocamIcon from '@mui/icons-material/Videocam';
 import EditIcon from '@mui/icons-material/Edit';
 import { meetingTimePipe } from '../utils/pipes';
 import Checkbox from '@mui/material/Checkbox';
-import { routes } from '../utils/routes';
-import { Link as RouterLink, useHistory } from 'react-router-dom';
-import { ChangeEvent, useState } from 'react';
-import { useAuth } from '../hooks/auth.hooks';
-import { useCreateStageGateChangeRequest } from '../hooks/change-requests.hooks';
-import { useToast } from '../hooks/toasts.hooks';
-import StageGateWorkPackageModal from './WorkPackageDetailPage/StageGateWorkPackageModalContainer/StageGateWorkPackageModal';
-
-// component for regular DR pills
-const Pill: React.FC<{
-  icon: React.ReactNode;
-  text: string;
-}> = ({ icon, text }) => {
-  return (
-    <Typography
-      sx={{
-        fontSize: 13,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        paddingX: 1,
-        width: 'fit-content'
-      }}
-    >
-      {icon}
-      <Typography paddingLeft="5px">{text}</Typography>
-    </Typography>
-  );
-};
-
-// component for the DR pills that are links (zoom and docs)
-const LinkPill: React.FC<{
-  icon: React.ReactNode;
-  linkText: string;
-  displayText: string;
-}> = ({ icon, linkText, displayText }) => {
-  return (
-    <Typography
-      sx={{
-        fontSize: 13,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        paddingX: 1,
-        width: 'fit-content'
-      }}
-    >
-      {icon}
-      <Link target="_blank" sx={{ color: 'white' }} href={linkText} paddingLeft="5px">
-        {displayText}
-      </Link>
-    </Typography>
-  );
-};
-
-// component for the member pill displaying full name
-const MemberPill: React.FC<{ user: User }> = ({ user }) => {
-  return (
-    <Box sx={{ borderRadius: '11px', backgroundColor: '#d9d9d9', width: 'fit-content', margin: '8px' }}>
-      <Typography
-        fontSize="15px"
-        paddingX="10px"
-        paddingY="1px"
-        color="#242526"
-      >{`${user.firstName} ${user.lastName}`}</Typography>
-    </Box>
-  );
-};
-
-interface StageGateModalProps {
-  wbsNum: WbsNumber;
-  modalShow: boolean;
-  handleClose: () => void;
-}
-
-export interface FormInput {
-  confirmDone: boolean;
-}
-
-// stage gate modal component (redacted loading and error for this specific case)
-const StageGateModal: React.FC<StageGateModalProps> = ({ wbsNum, modalShow, handleClose }) => {
-  const auth = useAuth();
-  const history = useHistory();
-  const toast = useToast();
-  const { mutateAsync } = useCreateStageGateChangeRequest();
-
-  const handleConfirm = async ({ confirmDone }: FormInput) => {
-    handleClose();
-    if (auth.user?.userId === undefined) throw new Error('Cannot create stage gate change request without being logged in');
-    try {
-      await mutateAsync({
-        submitterId: auth.user?.userId,
-        wbsNum,
-        type: ChangeRequestType.StageGate,
-        confirmDone
-      });
-      history.push(routes.CHANGE_REQUESTS);
-    } catch (e: unknown) {
-      if (e instanceof Error) {
-        toast.error(e.message);
-      }
-    }
-  };
-
-  return <StageGateWorkPackageModal wbsNum={wbsNum} modalShow={modalShow} onHide={handleClose} onSubmit={handleConfirm} />;
-};
-
-// delay wp modal component
-const DelayModal: React.FC<{ open: boolean; onHide: () => void; dr: DesignReview }> = ({ open, onHide, dr }) => {
-  const toast = useToast();
-  const [weeks, setWeeks] = useState<string>('');
-  const onChange = (e: ChangeEvent<HTMLInputElement>) => {
-    if (e.target.value === '' || parseInt(e.target.value) >= 1) {
-      setWeeks(e.target.value);
-    } else {
-      toast.error('If delaying, it must be by at least 1 week');
-    }
-  };
-
-  return (
-    <NERModal open={open} title="Delay WP" onHide={onHide} hideFormButtons showCloseButton>
-      <TextField
-        type="number"
-        label="Enter number of weeks"
-        variant="outlined"
-        value={weeks}
-        onChange={onChange}
-        fullWidth
-        margin="normal"
-      />
-      <Button
-        sx={{
-          color: 'white',
-          backgroundColor: '#EF4345',
-          ':hover': { backgroundColor: '#A72D2D' },
-          fontSize: 10,
-          fontWeight: 'bold',
-          marginLeft: '75%',
-          marginTop: '10px'
-        }}
-      >
-        <Link
-          underline="none"
-          color={'text.primary'}
-          component={RouterLink}
-          to={`${routes.CHANGE_REQUESTS}/new?wbsNum=${wbsPipe(dr.wbsNum)}&timelineDelay=${weeks}`}
-        >
-          Delay
-        </Link>
-      </Button>
-    </NERModal>
-  );
-};
+import { useState } from 'react';
+import { StageGateModal } from '../components/DesignReviewStageGateModal';
+import { DelayModal } from '../components/DesignReviewDelayModal';
+import { Pill } from '../components/DesignReviewPill';
+import { MemberPill } from '../components/DesignReviewMemberPill';
 
 interface DRCSummaryModalProps {
   open: boolean;
@@ -187,20 +39,22 @@ const DRCSummaryModal: React.FC<DRCSummaryModalProps> = ({ open, onHide, designR
         <Grid item>
           <Grid container direction="row" alignItems="center" justifyContent="center" columnSpacing={1}>
             <Grid item xs={3}>
-              <Pill icon={<AccessTimeIcon />} text={designReview.meetingTimes.map(meetingTimePipe).join(', ')} />
+              <Pill icon={<AccessTimeIcon />} isLink={false} displayText={meetingTimePipe(designReview.meetingTimes[0])} />
             </Grid>
             <Grid item xs={3}>
-              <Pill icon={<LocationOnIcon />} text={designReview.location ?? 'Online'} />
+              <Pill icon={<LocationOnIcon />} isLink={false} displayText={designReview.location ?? 'Online'} />
             </Grid>
             <Grid item xs={3}>
-              <LinkPill
+              <Pill
+                isLink
                 icon={<DescriptionIcon />}
                 linkText={designReview.docTemplateLink ?? ''}
                 displayText={designReview.docTemplateLink ? 'docs' : 'No Doc'}
               />
             </Grid>
             <Grid item xs={3}>
-              <LinkPill
+              <Pill
+                isLink
                 icon={<VideocamIcon />}
                 linkText={designReview.zoomLink ?? ''}
                 displayText={designReview.zoomLink ? 'zoom.us' : 'No Zoom'}
@@ -209,7 +63,7 @@ const DRCSummaryModal: React.FC<DRCSummaryModalProps> = ({ open, onHide, designR
           </Grid>
         </Grid>
         <Grid item xs={12}>
-          <Grid container direction="row" paddingY="40px">
+          <Grid container direction="row" paddingY="20px">
             <Grid item xs={12}>
               <Grid container>
                 <Grid item sx={{ display: 'flex', alignItems: 'start', marginTop: '7px' }}>
@@ -279,6 +133,7 @@ const DRCSummaryModal: React.FC<DRCSummaryModalProps> = ({ open, onHide, designR
                 marginLeft: 1,
                 fontWeight: 'bold'
               }}
+              disabled={designReview.status !== DesignReviewStatus.DONE}
             >
               Schedule Another DR
             </Button>
@@ -292,6 +147,7 @@ const DRCSummaryModal: React.FC<DRCSummaryModalProps> = ({ open, onHide, designR
                 fontWeight: 'bold'
               }}
               onClick={() => setShowStageGateModal(true)}
+              disabled={designReview.status !== DesignReviewStatus.DONE}
             >
               Stage Gate
             </Button>
@@ -305,6 +161,7 @@ const DRCSummaryModal: React.FC<DRCSummaryModalProps> = ({ open, onHide, designR
                 fontWeight: 'bold'
               }}
               onClick={() => setShowDelayModal(true)}
+              disabled={designReview.status !== DesignReviewStatus.DONE}
             >
               Request Delay to WP
             </Button>
